[LifetimeSafety] Refactor the analysis into smaller files
