+++ clang/include/clang/Analysis/Analyses/LifetimeSafety/LifetimeAnnotations.h
